requirements.txt
absl-py==2.1.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
asttokens==3.0.0
backcall==0.2.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
decord==0.6.0
deepspeed==0.5.8
einops==0.8.0
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.14.0
fonttools==4.55.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.68.0
hjson==3.1.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
openxlab==0.1.1
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
segment-anything==1.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
timm==0.4.12
torch==2.4.0+cu124
torch-tb-profiler==0.4.3
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
